WIP: Refactor rpc code both server and client
