+++ karte_flip_client/src/enums/routes_enums/SingletonAutoLoadEnum.cs
@@ -1,10 +1,10 @@
 public enum SingletonAutoLoadEnum{
     NETWORKING_SERVICE,
-    TURN_RPC_SERVICE,
     
     //Managers
     ROUTE_MANAGER,
     SCORING_MANAGER,
     PLAYER_MANAGER,
-    TURN_COMPUTER_MANAGER
+    COMPUTER_TURN_MANAGER,
+    PLAYER_TURN_MANAGER,
 }
