+++ karte_flip_server/project.godot
@@ -13,11 +13,11 @@ config_version=5
 config/name="karte_flip_server"
 run/main_scene="res://src/resources/scenes/server_scene.tscn"
 config/features=PackedStringArray("4.2", "C#", "Forward Plus")
-config/icon="res://icon.svg"
+config/icon="res://src/resources/assets/icons/icon.svg"
 
 [autoload]
 
-TURN_RPC_SERVICE="*res://src/services/TurnRpcService.cs"
+PLAYER_TURN_MANAGER="*res://src/managers/PlayerTurnManager.cs"
 
 [dotnet]
