+++ karte_flip_server/src/resources/assets/icons/icon.svg.import
@@ -3,15 +3,15 @@
 importer="texture"
 type="CompressedTexture2D"
 uid="uid://dso68wfw53p65"
-path="res://.godot/imported/icon.svg-218a8f2b3041327d8a5756f3a245f83b.ctex"
+path="res://.godot/imported/icon.svg-e53d9219aea45f5fc019d79ec864be0b.ctex"
 metadata={
 "vram_texture": false
 }
 
 [deps]
 
-source_file="res://icon.svg"
-dest_files=["res://.godot/imported/icon.svg-218a8f2b3041327d8a5756f3a245f83b.ctex"]
+source_file="res://src/resources/assets/icons/icon.svg"
+dest_files=["res://.godot/imported/icon.svg-e53d9219aea45f5fc019d79ec864be0b.ctex"]
 
 [params]
